# packages/api-client/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/workspaces": {
      "post": {
        "operationId": "WorkspacesController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkspaceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "supabase-auth": []
          }
        ],
        "summary": "Create a new personal workspace",
        "tags": [
          "Workspaces"
        ]
      },
      "get": {
        "operationId": "WorkspacesController_list_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkspaceResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "supabase-auth": []
          }
        ],
        "summary": "List user workspaces",
        "tags": [
          "Workspaces"
        ]
      }
    },
    "/v1/workspaces/{workspaceId}/summary": {
      "get": {
        "description": "Single aggregated payload (DEC-011). Amounts are minor units. Percentage deltas are NOT returned: with a zero prior month a change is undefined rather than +100%, so the client decides how to render that.",
        "operationId": "WorkspacesController_summary_v1",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceSummaryResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "supabase-auth": []
          }
        ],
        "summary": "Dashboard summary — net worth, month income/expense/net, prior month",
        "tags": [
          "Workspaces"
        ]
      }
    },
...
  },
  "info": {
    "title": "NoorixFin API",
    "description": "Personal and household finance management API. All financial amounts are minor-unit decimal strings. All timestamps are ISO 8601 UTC. Every failure body matches the `ApiErrorBody` schema; branch on its `code`, never on `message`.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "supabase-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Supabase access token"
      }
    },
    "schemas": {
...
      "CreateWorkspaceDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "My Personal Finance"
          },
          "base_currency": {
            "type": "string",
            "example": "BDT"
          },
          "timezone": {
            "type": "string",
            "example": "Asia/Dhaka"
          }
        },
        "required": [
          "name"
        ]
      },
      "WorkspaceResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "PERSONAL"
            ]
          },
          "name": {
            "type": "string"
          },
          "base_currency": {
            "type": "string"
          },
          "timezone": {
            "type": "string"
          },
          "created_by": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "PENDING_DELETION",
              "DELETED"
            ]
          },
          "created_at": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "description": "Current user role (always OWNER)"
          }
        },
        "required": [
          "id",
          "type",
          "name",
          "base_currency",
          "timezone",
          "created_by",
          "status",
          "created_at",
          "updated_at"
        ]
      },
      "WorkspaceSummaryResponseDto": {
        "type": "object",
        "properties": {
          "visible": {
            "type": "boolean"
          },
          "net_worth": {
            "type": "number"
          },
          "income": {
            "type": "number"
          },
          "expense": {
            "type": "number"
          },
          "net": {
            "type": "number"
          },
          "prev_income": {
            "type": "number"
          },
          "prev_expense": {
            "type": "number"
          },
          "prev_net": {
            "type": "number"
          },
          "account_count": {
            "type": "number"
          },
          "timezone": {
            "type": "string"
          },
          "month_start": {
            "type": "string"
          }
        },
        "required": [
          "visible",
          "net_worth",
          "income",
          "expense",
          "net",
          "prev_income",
          "prev_expense",
          "prev_net",
          "account_count"
        ]
      },
...
    }
  }
}